fix and refactor: refined the interactive mode and code structure
